backend/french_version/requirements.txt
Flask
psycopg2-binary
requests
gunicorn
Werkzeug
MarkupSafe
flask-cors